Fixed https://sonarcloud.io/project/issues?assignees=TimZue%40github&resolved=false&id=bettercodepaul_swt2-bsa-backend&open=AX7dl2o1aaBp2JXSJ_5S, by renaming the package to its current path
