+++ bogenliga/bogenliga-application/src/main/java/de/bogenliga/application/services/v1/sportjahr/mapper/SportjahrDTOMapper.java
@@ -1,7 +1,7 @@
 package de.bogenliga.application.services.v1.sportjahr.mapper;
 import de.bogenliga.application.business.sportjahr.api.types.SportjahrDO;
 import de.bogenliga.application.common.service.mapping.DataTransferObjectMapper;
-import de.bogenliga.application.services.v1.sportjahr.SportjahrDTO;
+import de.bogenliga.application.services.v1.sportjahr.model.SportjahrDTO;
 import java.util.function.*;
 
 /**
